+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## ???
+
+### Fixes
+
+* Fix: starting with Peertube 3.2.0, there is a header 'X-Frame-Options'. Removing it on the iframe route.
+
 ## v2.1.3
 
 * Fix: 2.1.0 was in fact correct... Did not work on my preprod env because of... a Livebox bug...