+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-* Don't display webchat when viewing a playlist.
+* Don't display webchat when viewing a playlist (requires Peertube 3.2.0).
 * Better default values for settings.
 
 ### Fixes