+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## ???
+## v2.2.0
 
 ### Features